requirements.txt
langchain==0.1.12
langchain-community==0.0.27
streamlit==1.32.2
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.0
scikit-learn==1.4.1
numpy==1.26.4
requests==2.31.0 